backend/requirements.txt
django-celery-beat==2.6.0
django==4.2
djangorestframework==3.14.0
django-filter==23.2
drf-yasg==1.21.5
django-simple-history==3.7.0
django-import-export==3.2.0
django-redis==5.4.0
pylint==3.2.3
django-cors-headers==4.3.1